Split composer prompt input panel
