+++ src/app/components/workspace/composer/useComposerSlashCommands.ts
@@ -62,6 +62,8 @@ type UseComposerSlashCommandsOptions = {
   onOpenSettingsView: () => void;
 };
 
+export type ComposerSlashCommands = ReturnType<typeof useComposerSlashCommands>;
+
 export function useComposerSlashCommands({
   draft,
   projectId,
